+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleResourceInterpreter.scala
@@ -5,7 +5,7 @@ import cats.effect.concurrent.Semaphore
 import cats.effect.{Async, Blocker, ContextShift, Timer}
 import cats.mtl.Ask
 import com.google.cloud.resourcemanager.{Project, ResourceManager}
-import io.chrisdavenport.log4cats.StructuredLogger
+import org.typelevel.log4cats.StructuredLogger
 import org.broadinstitute.dsde.workbench.model.TraceId
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 import cats.syntax.all._