+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleSubscriptionAdmin.scala
@@ -9,7 +9,7 @@ import com.google.auth.oauth2.ServiceAccountCredentials
 import com.google.cloud.pubsub.v1.{SubscriptionAdminClient, SubscriptionAdminSettings}
 import com.google.pubsub.v1.{ProjectSubscriptionName, Subscription}
 import fs2.Stream
-import io.chrisdavenport.log4cats.StructuredLogger
+import org.typelevel.log4cats.StructuredLogger
 import org.broadinstitute.dsde.workbench.model.TraceId
 import org.broadinstitute.dsde.workbench.model.google.GoogleProject