+++ project/Dependencies.scala
@@ -8,7 +8,7 @@ object Dependencies {
   val scalaLoggingV = "3.9.3"
   val scalaTestV    = "3.2.6"
   val circeVersion = "0.13.0"
-  val http4sVersion = "0.21.19"
+  val http4sVersion = "0.21.20"
 
   def excludeGuavaJDK5(m: ModuleID): ModuleID = m.exclude("com.google.guava", "guava-jdk5")
 
@@ -75,7 +75,7 @@ object Dependencies {
   val circeParser: ModuleID = "io.circe" %% "circe-parser" % circeVersion
   val circeGeneric: ModuleID = "io.circe" %% "circe-generic" % circeVersion % "test"
   val circeFs2: ModuleID = "io.circe" %% "circe-fs2" % circeVersion
-  val log4cats = "io.chrisdavenport" %% "log4cats-slf4j"   % "1.1.1"
+  val log4cats = "org.typelevel" %% "log4cats-slf4j"   % "1.2.0"
   val catsMtl = "org.typelevel" %% "cats-mtl" % "1.1.2"
 
   val http4sCirce = "org.http4s" %% "http4s-circe" % http4sVersion